Temporary partial fix to documentation lookup.
We currently fail to find documentation in the face of re-exports. This
is most painful with definitions provided by base, as GHC's internal
module names are different than those imported by the user. This is a
temporary measure to resolve such issues (e.g. an identifier defined in
"GHC.List" has its documentation in "Data.List").